# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Information Crawler API",
    "summary": "中关村人工智能研究院 — 信息监测系统",
    "description": "## 概述\n\n信息监测系统 API，自动爬取 **129 个信源**（105 个启用），覆盖 **9 个维度**，为中关村人工智能研究院提供全方位的信息监测与商业智能服务。\n\n## 功能模块\n\n| 模块 | 说明 |\n|------|------|\n| **文章管理** | 全量文章的查询、搜索、统计 |\n| **信源管理** | 129 个信源的配置、状态监控、手动触发 |\n| **维度视图** | 9 大维度的文章聚合浏览 |\n| **政策智能** | 规则引擎 + LLM 二级管线，政策机会挖掘 |\n| **人事情报** | 任免信息自动提取，LLM 相关性分析 |\n| **系统健康** | 数据库、调度器、爬取健康度监控 |\n\n## 维度说明\n\n- `national_policy` — 国家政策（国务院、部委）\n- `beijing_policy` — 北京政策（市/区政府）\n- `technology` — 技术动态（ArXiv、GitHub Trending、Hacker News 等）\n- `talent` — 人才政策\n- `industry` — 产业动态\n- `universities` — 高校动态（46 所高校 AI 院系）\n- `events` — 活动会议\n- `personnel` — 人事变动\n- `twitter` — Twitter/X KOL 动态\n\n## 技术栈\n\nFastAPI + SQLAlchemy(async) + PostgreSQL(Supabase) + APScheduler + httpx + BeautifulSoup4 + Playwright",
    "contact": {
      "name": "中关村人工智能研究院",
      "url": "https://www.zgcaiia.com/"
    },
    "license": {
      "name": "Internal Use"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/articles/": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "文章列表",
        "description": "查询文章列表，支持按维度、信源、关键词、日期范围过滤，以及字段排序和分页。",
        "operationId": "list_articles_api_v1_articles__get",
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dimension",
              "title": "Dimension"
            },
            "description": "Filter by dimension"
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source ID",
              "title": "Source Id"
            },
            "description": "Filter by source ID"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyword filter in title/content",
              "title": "Keyword"
            },
            "description": "Keyword filter in title/content"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (ISO format)",
              "title": "Date From"
            },
            "description": "Start date (ISO format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (ISO format)",
              "title": "Date To"
            },
            "description": "End date (ISO format)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "crawled_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order: asc or desc",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort order: asc or desc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ArticleBrief_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/search": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "全文搜索",
        "description": "在文章标题和正文中进行关键词搜索，参数与列表接口相同。",
        "operationId": "search_articles_api_v1_articles_search_get",
        "parameters": [
          {
            "name": "dimension",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dimension",
              "title": "Dimension"
            },
            "description": "Filter by dimension"
          },
          {
            "name": "source_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source ID",
              "title": "Source Id"
            },
            "description": "Filter by source ID"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyword filter in title/content",
              "title": "Keyword"
            },
            "description": "Keyword filter in title/content"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (ISO format)",
              "title": "Date From"
            },
            "description": "Start date (ISO format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (ISO format)",
              "title": "Date To"
            },
            "description": "End date (ISO format)"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "crawled_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order: asc or desc",
              "default": "desc",
              "title": "Order"
            },
            "description": "Sort order: asc or desc"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ArticleBrief_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/stats": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "文章统计",
        "description": "获取文章聚合统计，默认按维度分组。可选按信源分组。",
        "operationId": "get_stats_api_v1_articles_stats_get",
        "parameters": [
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dimension",
              "title": "Group By"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleStats"
                  },
                  "title": "Response Get Stats Api V1 Articles Stats Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{article_id}": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "文章详情",
        "description": "根据文章 ID 获取完整文章内容，包含正文和额外元数据。",
        "operationId": "get_article_api_v1_articles__article_id__get",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Article Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetail"
                }
              }
            }
          },
          "404": {
            "description": "文章不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "articles"
        ],
        "summary": "更新文章",
        "description": "更新文章元数据，如标记已读状态或设置重要度评分。",
        "operationId": "update_article_api_v1_articles__article_id__patch",
        "parameters": [
          {
            "name": "article_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Article Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDetail"
                }
              }
            }
          },
          "404": {
            "description": "文章不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "文章唯一 ID",
            "examples": [
              1024
            ]
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "description": "所属信源 ID",
            "examples": [
              "tech_arxiv"
            ]
          },
          "dimension": {
            "type": "string",
            "title": "Dimension",
            "description": "所属维度",
            "examples": [
              "technology"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "文章原始链接",
            "examples": [
              "https://arxiv.org/abs/2401.12345"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "文章标题",
            "examples": [
              "Large Language Models for Code Generation: A Survey"
            ]
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "作者",
            "examples": [
              "张三"
            ]
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At",
            "description": "发布时间（ISO 8601）"
          },
          "crawled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Crawled At",
            "description": "爬取时间（ISO 8601）"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "标签列表",
            "default": [],
            "examples": [
              [
                "AI",
                "LLM",
                "政策"
              ]
            ]
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read",
            "description": "是否已读",
            "default": false
          },
          "importance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance",
            "description": "重要度评分（0-100）",
            "examples": [
              85
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "dimension",
          "url",
          "title",
          "crawled_at"
        ],
        "title": "ArticleBrief",
        "description": "文章摘要信息，用于列表展示。"
      },
      "ArticleDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "文章唯一 ID",
            "examples": [
              1024
            ]
          },
          "source_id": {
            "type": "string",
            "title": "Source Id",
            "description": "所属信源 ID",
            "examples": [
              "tech_arxiv"
            ]
          },
          "dimension": {
            "type": "string",
            "title": "Dimension",
            "description": "所属维度",
            "examples": [
              "technology"
            ]
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "文章原始链接",
            "examples": [
              "https://arxiv.org/abs/2401.12345"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "文章标题",
            "examples": [
              "Large Language Models for Code Generation: A Survey"
            ]
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "作者",
            "examples": [
              "张三"
            ]
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At",
            "description": "发布时间（ISO 8601）"
          },
          "crawled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Crawled At",
            "description": "爬取时间（ISO 8601）"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "标签列表",
            "default": [],
            "examples": [
              [
                "AI",
                "LLM",
                "政策"
              ]
            ]
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read",
            "description": "是否已读",
            "default": false
          },
          "importance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance",
            "description": "重要度评分（0-100）",
            "examples": [
              85
            ]
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "文章正文（HTML 或纯文本）"
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra",
            "description": "额外元数据（JSON），不同信源可能包含不同字段",
            "default": {},
            "examples": [
              {
                "pdf_url": "https://arxiv.org/pdf/2401.12345"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_id",
          "dimension",
          "url",
          "title",
          "crawled_at"
        ],
        "title": "ArticleDetail",
        "description": "文章详情，包含正文内容和额外字段。"
      },
      "ArticleStats": {
        "properties": {
          "group": {
            "type": "string",
            "title": "Group",
            "description": "分组名称（维度名或信源 ID）",
            "examples": [
              "technology"
            ]
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "该分组下的文章总数",
            "examples": [
              342
            ]
          }
        },
        "type": "object",
        "required": [
          "group",
          "count"
        ],
        "title": "ArticleStats",
        "description": "文章统计信息。"
      },
      "ArticleUpdate": {
        "properties": {
          "is_read": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Read",
            "description": "标记为已读/未读"
          },
          "importance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Importance",
            "description": "设置重要度评分（0-100）",
            "examples": [
              90
            ]
          }
        },
        "type": "object",
        "title": "ArticleUpdate",
        "description": "文章更新请求体。"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "错误详情",
            "examples": [
              "Article not found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaginatedResponse_ArticleBrief_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArticleBrief"
            },
            "type": "array",
            "title": "Items",
            "description": "当前页的数据列表"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "符合条件的总记录数",
            "examples": [
              156
            ]
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "当前页码（从 1 开始）",
            "examples": [
              1
            ]
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "每页条数",
            "examples": [
              20
            ]
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "总页数",
            "examples": [
              8
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PaginatedResponse[ArticleBrief]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}